requirements.txt
streamlit==1.39.0
pandas==2.5.1
scikit-learn==1.5.0
nltk==3.9.1
Pillow==10.5.0
pytesseract==0.3.12
PyPDF2==3.0.1
python-docx==1.1.2
fastapi
uvicorn
